+++ src/org/schema/enumerations/status/ReservationCancelled.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace org\schema\enumerations\status;
+
+/**
+ * The status for a previously confirmed reservation that is now cancelled.
+ * @see https://schema.org/ReservationCancelled
+ */
+class ReservationCancelled extends ReservationStatusType
+{
+
+}