+++ src/org/schema/enumerations/status/ReservationConfirmed.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace org\schema\enumerations\status;
+
+/**
+ * The status of a confirmed reservation.
+ * @see https://schema.org/ReservationConfirmed
+ */
+class ReservationConfirmed extends ReservationStatusType
+{
+
+}