+++ src/org/schema/enumerations/status/ReservationHold.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace org\schema\enumerations\status;
+
+/**
+ * The status of a reservation on hold pending an update like credit card number or flight changes.
+ * @see https://schema.org/ReservationHold
+ */
+class ReservationHold extends ReservationStatusType
+{
+
+}