+++ src/org/schema/enumerations/status/ReservationPending.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace org\schema\enumerations\status;
+
+/**
+ * The status of a reservation when a request has been sent, but not confirmed.
+ * @see https://schema.org/ReservationPending
+ */
+class ReservationPending extends ReservationStatusType
+{
+
+}